BonjourServiceをモダンなNetwork frameworkで完全書き直し
NSNetServiceからNWListenerベースの実装に完全移行:

## 変更点
- Network frameworkのNWListenerを使用
- NWTXTRecordで型安全なTXTレコード管理
- iOS 13以降の最新APIを採用
- 非同期状態管理の改善
- エラーハンドリングの強化

## 技術的改善
- deprecatedなAPIを完全削除
- モダンなSwift concurrency対応
- より信頼性の高いサービス発信
- TXTレコード更新の簡素化

これで最新のiOSバージョンに最適化されたBonjourサービスになりました。

+++ ReuseBackupServer/ReuseBackupServerTests/Services/BonjourServiceTests.swift
@@ -8,7 +8,7 @@ struct BonjourServiceTests {
     // MARK: - Initialization Tests
     
     @Test func when_bonjour_service_initialized_then_properties_are_set_correctly() async throws {
-        let port: Int32 = 8080
+        let port: UInt16 = 8080
         let bonjourService = BonjourService(port: port)
         
         // 初期状態の確認
